*** empty log message ***

+++ test/userentry.cpp
@@ -74,7 +74,7 @@ int main(int argc, const char **argv) {
 	test("sid",uent.getSidString() &&
 			charstring::compare(uent.getSidString(),"-1"));
 #else
-	char	*uidstr=charstring::parseNumber(uent.getUserId());
+	char	*uidstr=charstring::parseNumber((uint64_t)uent.getUserId());
 	test("sid",uent.getSidString() &&
 			!charstring::compare(uidstr,uent.getSidString()));
 	delete[] uidstr;
